Provide all airspace profile segments intersecting an interval of interest

Provide all airspace profile segments intersecting an interval of interest

Usage

airspace_profiles_tidy(wef, til, airspace = "FIR", profile = "CTFM")

Arguments

wef

With EFfect date (included) at Zulu time in a format recognized by as_datetime()

til

TILl date (excluded) at Zulu time in a format recognized by as_datetime()

airspace

the type of airspace (default: 'FIR'), one of:

'FIR' (Flight Information Region)

'NAS' (National Airspace)

'AUA' (ATC Unit Airspace)

'ES' (Elementary Sector)

profile

the model of the trajectory profile (default: 'CTFM'), one of:

'FTFM'

'RTFM'

'CTFM'

'CPF'

'DCT'

'SCR'

'SRR'

'SUR'

Value

a tbl with the following columns

ID: the so called SAM ID, used internally by PRISME

SEQ_ID: the sequence number of the segment for the relevant airspace profile

ENTRY_TIME: the time of entry into the relevant airspace

ENTRY_LON: the longitude of entry into the relevant airspace

ENTRY_LAT: the latitude of entry into the relevant airspace

ENTRY_FL: the flight level of entry into the relevant airspace

EXIT_TIME: the time of exit out of the relevant airspace

EXIT_LON: the longitude of exit out of the relevant airspace

EXIT_LAT: the latitude of exit out of the relevant airspace

EXIT_FL: the flight level of exit out of the relevant airspace

AIRSPACE_ID: the airspace ID

AIRSPACE_TYPE: the airspace type as per airspace input parameter

MODEL_TYPE: the trajectory model as per profile input parameter

Examples

if (FALSE) {
asp_profs <- airspace_profiles_tidy(wef = "2023-01-01", til = "2023-04-01")
}