Extract a clean flights list in an interval

The returned tbl includes scheduled and non-scheduled flight departing in the right-opened interval [wef, til). General aviation, State, military and sensitive flight are excluded.

NOTE: you need access to PRU_DEV schema

Usage

flights_tidy(con = NULL, wef, til)

Arguments

con

DB connection to be used, if NULL instantiate one.

wef

With EFfect date (included) at Zulu time in a format recognized by lubridate::as_datetime()

til

uTILl date (excluded) at Zulu time in a format recognized by lubridate::as_datetime()

Value

a tbl with the following columns

FLT_UID: the flight unique id

LOBT: Last received Off-Block Time

IOBT: Initial Off-Block Time

AIRCRAFT_ID

REGISTRATION: the registration

AIRCRAFT_TYPE_ICAO_ID

FLT_RULES (see FPL Item 8): which sets of regulations the flight is operated under. Possible values are:

I for IFR

V for VFR,

Y first IFR thereafter VFR,

Z first VFR thereafter IFR

ICAO_FLT_TYPE (see FPL Item 8): flight type. Possible values (but G, M aand X shouldn't appear):

S for scheduled air service

N for non-scheduled air service

G for general aviation

M for military

X for other than the preceding categories

WK_TBL_CAT (see FPL Item 9): wake turbulence category, cab be

L LIGHT, i.e. maximum certificated takeoff mass of 7000 kg (15_500 lbs) or less.

M MEDIUM, i.e maximum certificated takeoff mass less than 136_000 kg (300_000 lbs), but more than 7_000 kg (15_500 lbs)

H HEAVY, i.e. maximum certificated takeoff mass of 136_000 kg (300_000 lbs) or more (except those specified as J)

J SUPER, presently the only the AIRBUS A-380-800

AIRCRAFT_OPERATOR:

AIRCRAFT_ADDRESS: the ICAO 24-bit address of the airframe for ADS-B/Mode S broadcasting

ADEP: (ICAO code of the) Aerodrome of DEParture

ADES: (ICAO code of the) Aerodrome of DEStination

ID: the so called SAM ID, used internally by PRISME

EOBT_1: Estimated Off-Block Time for FPL-based (M1) trajectory

ARVT_1: ARriVal Time for FPL-based (M1) trajectory

TAXI_TIME_1: Taxi time for FPL-based (M1) trajectory

AOBT_3: Actual Off-Block Time for flown (M3) trajectory

ARVT_3: ARVival Time for flown (M3) trajectory

TAXI_TIME_3: Taxi time for flown (M3) trajectory

RULE_NAME: market segment type

AO_GRP_CODE: Aircraft Operator group (code)

AO_GRP_NAME: : Aircraft Operator group (name)

RTE_LEN_1: route length (NM) for FPL-based (M1) trajectory

RTE_LEN_3: route length (NM) for for flown (M3) trajectory

Examples

if (FALSE) {
my_flts <- flights_tidy(wef = "2023-01-01", til = "2023-04-01")
}