Extract a clean airport operator data flow list in an interval

The returned dbplyr::tbl_dbi() includes movements information in the interval [wef, til). NOTE: it can only cover ONE month at a time

Usage

aodf_tidy(conn = NULL, wef, til)

Arguments

conn

Database connection or instantiate the default one.

wef

With EFfect date (included) at Zulu time in a format recognized by lubridate::as_datetime()

til

unTILl date (excluded) at Zulu time in a format recognized by lubridate::as_datetime()

Value

A dbplyr::tbl_dbi() with the following columns:

APDS_ID: the airport operator dataflow unique record id.

ID: the so called SAM ID, used internally by PRISME

AP_C_FLTID: flight identifier (aource Airport)

AP_C_FLTRUL: which sets of regulations the flight is operated under. Possible values are:

IFR for IFR

VFR for VFR

NA if unknown

AP_C_REG: the aircraft registration (with spaces, dashes, ... stripped), e.g. GEUUU.

ADEP_ICAO: (ICAO code of the) Aerodrome of DEParture (source airport).

ADES_ICAO: (ICAO code of the) Aerodrome of DEStination (source airport).

SRC_PHASE: flight phase. DEP=departure, ARR=arrival.

MVT_TIME_UTC: (best available) movement time (takeoff if SRC_PHASE = DEP, landing if SRC_PHASE = ARR).

BLOCK_TIME_UTC: Block time (off-block if SRC_PHASE = DEP, in-block if SRC_PHASE = ARR).

SCHED_TIME_UTC: scheduled time (of departure if SRC_PHASE = DEP, of arrival if SRC_PHASE = ARR; source airport).

ARCTYP: (best available) the ICAO code for the aircraft type, for example A21N for Airbus A321neo.

AP_C_RWY: Runway ID (of departure if SRC_PHASE = DEP, of arrival if SRC_PHASE = ARR; source airport).

AP_C_STND: Stand ID (of departure if SRC_PHASE = DEP, of arrival if SRC_PHASE = ARR; source airport).

C40_CROSS_TIME: time of first (last) crossing at 40 NM from ARP for departure (arrival).

C40_CROSS_LAT: latitude of first (last) crossing at 40 NM from ARP for departure (arrival).

C40_CROSS_LON: longitude of first (last) crossing at 40 NM from ARP for departure (arrival).

C40_CROSS_FL: flight level of first (last) crossing at 40 NM from ARP for departure (arrival).

C40_BEARING: bearing of first (last) crossing at 40 NM from ARP for departure (arrival).

C100_CROSS_TIME: time of first (last) crossing at 100 NM from ARP for departure (arrival).

C100_CROSS_LAT: latitude of first (last) crossing at 100 NM from ARP for departure (arrival).

C100_CROSS_LON: longitude of first (last) crossing at 100 NM from ARP for departure (arrival).

C100_CROSS_FL: flight level of first (last) crossing at 100 NM from ARP for departure (arrival).

C100_BEARING: bearing of first (last) crossing at 100 NM from ARP for departure (arrival).

Note

You need to either provide a connection conn that has access to SWH_FCT.FAC_APDS_FLIGHT_IR691, or go with the default which uses PRU_ATMAP to establish a db_connection().

Examples

if (FALSE) { # \dontrun{
my_aodf <- aodf_tidy(wef = "2023-01-01", til = "2023-01-02")
# ...
DBI::dbDisconnect(my_aodf$src$con)
} # }